pyproject.toml
[project]
name = "unitn-mds"
version = "0.1.0"
description = "Multimedia Data Security Laboratory"
requires-python = "==3.8.10"
dependencies = [
    "common",
    "ipykernel",
    "matplotlib",
    "notebook",
    "numpy",
    "opencv-python",
    "pandas",
    "pycompile",
    "scikit-image",
    "scikit-learn",
    "scipy",
]
